tools/dsal/requirements.txt
jinja2
netaddr
openstacksdk